+++ src/main/java/com/witchworks/common/core/proxy/ISidedProxy.java
@@ -21,5 +21,5 @@ public interface ISidedProxy {
 
 	void displayRecordText(ITextComponent text);
 
-	void spawnParticle(ParticleF particleF, double x, double y, double z, double xSpeed, double ySpeed, double zSpeed, float... args);
+	void spawnParticle(ParticleF particleF, double x, double y, double z, double xSpeed, double ySpeed, double zSpeed, int... args);
 }
